# api-gateway/swagger.yaml
openapi: 3.0.1
info:
  title: book-recommendations
  description: book-recommendations
  version: v1
  contact:
    name: Julio Scheidt
    email: julio@blackdevs.com.br
servers:
  - url: http://book-recommendations.{NAMESPACE}.svc.cluster.local:3080/
...
paths:
...
  /v1/user/{uuid}:
    get:
      tags:
        - UserServices
      summary: get user
      operationId: getUser
      security:
        - BearerAuth: []
      parameters:
      - name: uuid
        in: path
        description: Uuid for user
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
      responses:
        '200':
          description: 'OK'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetUserSuccessResponse'
        '400':
          description: 'Bad Request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseBadRequest'
        '401':
          description: 'Unauthorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseUnauthorized'
        '404':
          description: 'Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseNotFound'
        '500':
          description: 'Internal Server Error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseServerError'

    put:
      tags:
        - UserServices
      summary: update user
      operationId: updateUser
      security:
        - BearerAuth: []
      parameters:
      - name: uuid
        in: path
        description: Uuid for user
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/UserBody'
                - type: object
                  properties:
                    password:
                      type: string
      responses:
        '202':
          description: 'Accepted'
        '400':
          description: 'Bad Request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseBadRequest'
        '401':
          description: 'Unauthorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseUnauthorized'
        '404':
          description: 'Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseNotFound'
        '500':
          description: 'Internal Server Error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseServerError'

    delete:
      tags:
        - UserServices
      summary: delete user
      operationId: deleteUser
      security:
        - BearerAuth: []
      parameters:
      - name: uuid
        in: path
        description: Uuid for user
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
      responses:
        '204':
          description: 'No Content'
        '400':
          description: 'Bad Request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseBadRequest'
        '401':
          description: 'Unauthorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseUnauthorized'
        '404':
          description: 'Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseNotFound'
        '500':
          description: 'Internal Server Error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseServerError'
  /v1/user/rate:
    post:
      tags:
        - UserRateServices
      summary: upsert user rate
      operationId: upsertUserRate
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/UserRateBody'
                - type: object
                  properties:
                    book_uuid:
                      $ref: '#/components/schemas/Uuid'
                    rate:
                      type: number
      responses:
        '201':
          description: 'Created'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpsertUserRateSuccessResponse'
        '400':
          description: 'Bad Request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseBadRequest'
        '401':
          description: 'Unauthorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseUnauthorized'
        '404':
          description: 'Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseNotFound'
        '500':
          description: 'Internal Server Error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseServerError'

    get:
      tags:
        - UserRateServices
      summary: list user rate
      operationId: listUserRate
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page to search
          required: true
          schema:
            type: integer
            example: 0
            minimum: 0
        - name: size
          in: query
          description: Size of page
          required: true
          schema:
            type: integer
            example: 50
            maximum: 100
      responses:
        '200':
          description: 'OK'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListUserRateSuccessResponse'
        '401':
          description: 'Unauthorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseUnauthorized'
        '500':
          description: 'Internal Server Error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseServerError'
  /v1/user/rate/{user_uuid}:
    get:
      tags:
        - UserRateServices
      summary: get user rate
      operationId: getUserRate
      security:
        - BearerAuth: []
      parameters:
      - name: user_uuid
        in: path
        description: Uuid for user
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
      responses:
        '200':
          description: 'OK'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetUserRateSuccessResponse'
        '400':
          description: 'Bad Request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseBadRequest'
        '401':
          description: 'Unauthorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseUnauthorized'
        '404':
          description: 'Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseNotFound'
        '500':
          description: 'Internal Server Error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseServerError'

    delete:
      tags:
        - UserRateServices
      summary: delete user rate
      operationId: deleteUserRate
      security:
        - BearerAuth: []
      parameters:
      - name: user_uuid
        in: path
        description: Uuid for user
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
      responses:
        '204':
          description: 'No Content'
        '400':
          description: 'Bad Request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseBadRequest'
        '401':
          description: 'Unauthorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseUnauthorized'
        '404':
          description: 'Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseNotFound'
        '500':
          description: 'Internal Server Error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseServerError'
...
  /v1/book/{uuid}:
    get:
      tags:
        - BookServices
      summary: get book
      operationId: getBook
      security:
        - BearerAuth: []
      parameters:
      - name: uuid
        in: path
        description: Uuid for book
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
      responses:
        '200':
          description: 'OK'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetBookSuccessResponse'
        '400':
          description: 'Bad Request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseBadRequest'
        '401':
          description: 'Unauthorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseUnauthorized'
        '404':
          description: 'Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseNotFound'
        '500':
          description: 'Internal Server Error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseServerError'

    put:
      tags:
        - BookServices
      summary: update book
      operationId: updateBook
      security:
        - BearerAuth: []
      parameters:
      - name: uuid
        in: path
        description: Uuid for book
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookBody'
      responses:
        '202':
          description: 'Accepted'
        '400':
          description: 'Bad Request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseBadRequest'
        '401':
          description: 'Unauthorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseUnauthorized'
        '404':
          description: 'Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseNotFound'
        '500':
          description: 'Internal Server Error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseServerError'

    delete:
      tags:
        - BookServices
      summary: delete book
      operationId: deleteBook
      security:
        - BearerAuth: []
      parameters:
      - name: uuid
        in: path
        description: Uuid for book
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
      responses:
        '204':
          description: 'No Content'
        '400':
          description: 'Bad Request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseBadRequest'
        '401':
          description: 'Unauthorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseUnauthorized'
        '404':
          description: 'Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseNotFound'
        '500':
          description: 'Internal Server Error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseServerError'
  /v1/recommendation/user/{uuid}:
    get:
      tags:
        - RecommendationServices
      summary: get recommendation
      operationId: getRecommendation
      security:
        - BearerAuth: []
      parameters:
      - name: uuid
        in: path
        description: Uuid for recommendation
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
      responses:
        '200':
          description: 'OK'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetRecommendationSuccessResponse'
        '400':
          description: 'Bad Request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseBadRequest'
        '401':
          description: 'Unauthorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseUnauthorized'
        '404':
          description: 'Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseNotFound'
        '500':
          description: 'Internal Server Error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseServerError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    DefaultResponseBadRequest:
      type: object
      properties:
        message:
          type: string
          description: 'Status code'
          example: 'Bad Request'
    DefaultResponseUnauthorized:
      type: object
      properties:
        message:
          type: string
          description: 'Status code'
          example: 'Unauthorized'
    DefaultResponseNotFound:
      type: object
      properties:
        message:
          type: string
          description: 'Status code'
          example: 'Not Found'
    DefaultResponseServerError:
      type: object
      properties:
        message:
          type: string
          description: 'Status code'
          example: 'Internal Server Error'
    Uuid:
      type: string
      example: '17f182b0dfc002737d507f74'
    CreatedAt:
      type: string
      format: date-time
      example: '2021-01-01T12:00:00.000Z'
    UpdatedAt:
      type: string
      format: date-time
      example: '2021-01-01T12:00:00.000Z'
    UserBody:
      type: object
      properties:
        name:
          type: string
        surname:
          type: string
        email:
          type: string
        phone:
          type: string
          format: phone
          example: '1199999999'
    UserFull:
      allOf:
        - type: object
          properties:
            uuid:
              $ref: '#/components/schemas/Uuid'
            created_at:
              $ref: '#/components/schemas/CreatedAt'
            updated_at:
              $ref: '#/components/schemas/UpdatedAt'
        - $ref: '#/components/schemas/UserBody'
    UserRateBody:
      allOf:
        - type: object
          properties:
            user_uuid:
              $ref: '#/components/schemas/Uuid'
    UserRateFull:
      allOf:
        - $ref: '#/components/schemas/UserRateBody'
        - type: object
          properties:
            rates:
              type: array
              items:
                type: object
                properties:
                  book_uuid:
                    $ref: '#/components/schemas/Uuid'
                  rate:
                    type: number
            created_at:
              $ref: '#/components/schemas/CreatedAt'
            updated_at:
              $ref: '#/components/schemas/UpdatedAt'
    BookBody:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        genre:
          type: string
        image:
          type: string
    BookFull:
      allOf:
        - type: object
          properties:
            uuid:
              $ref: '#/components/schemas/Uuid'
            created_at:
              $ref: '#/components/schemas/CreatedAt'
            updated_at:
              $ref: '#/components/schemas/UpdatedAt'
        - $ref: '#/components/schemas/BookBody'
    RecommendationBody:
      type: object
      properties:
        user_uuid:
          $ref: '#/components/schemas/Uuid'
        recommendations:
          type: array
          items:
            type: object
            properties:
              book_uuid:
                $ref: '#/components/schemas/Uuid'
              rate:
                type: number
    RecommendationFull:
      $ref: '#/components/schemas/RecommendationBody'
...
    GetUserSuccessResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/UserFull'
...
    UpsertUserRateSuccessResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Uuid'
    GetUserRateSuccessResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/UserRateFull'
    ListUserRateSuccessResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/UserRateFull'
...
    GetBookSuccessResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/BookFull'
...
    GetRecommendationSuccessResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/RecommendationFull'